refactor(cli): share runScriptCommand across dev and init handlers
Lift runScriptCommand into util/detect-package-manager.ts (alongside
execArgsFor/installArgsFor) and use it in both handlers: init imports it instead
of defining its own copy, and dev's frameworkDevScript (which was exactly
runScriptCommand(manager, "dev")) is replaced by that call. Same output; removes
the duplicated package-manager script-command spelling.
